% Generated by roxygen2 (4.0.2): do not edit by hand
\name{element_texts}
\alias{element_texts}
\title{Read element texts from an Omeka item}
\usage{
element_texts(item, filter)
}
\arguments{
\item{item}{An Omeka item as a list}

\item{filter}{Which element texts to return, as a character vector}
}
\value{
A named character vector of element texts
}
\description{
Omeka items as returned by the API have a deeply nested structure. Most of
the interesting information is stored in the element texts. This function
provides a convenient way of accessing those element texts. Calling this
function on an Omeka item returns a named character vector of all the element
texts. Optionally you can pass a character vector to filter the output so the
function returns only a particular kind of metadata, e.g., \code{"Title"}.
}
\examples{
omeka_endpoint("http://20.rrchnm.org/api")
# Get the Papers of the War Department project item
war_dept <- omeka_get("items", 197)
element_texts(war_dept)
element_texts(war_dept, "Title")
element_texts(war_dept, "Content Experts")
element_texts(war_dept, c("Start Date", "End Date"))
}

