% Generated by roxygen2 (4.0.2): do not edit by hand
\name{omeka_endpoint}
\alias{omeka_endpoint}
\title{Get or set the Omeka endpoint}
\usage{
omeka_endpoint(endpoint = NULL)
}
\arguments{
\item{endpoint}{The URL of the Omeka endpoint you are using}
}
\value{
The URL of the Omeka endpoint
}
\description{
Pass a URL to this function to set the Omeka API endpoint for the rest of
your script. Call this function without an argument to get the currently set
endpoint.
}
\examples{
omeka_endpoint("http://20.rrchnm.org/api")
omeka_endpoint()
}

