% Generated by roxygen2 (4.0.2): do not edit by hand
\name{omeka_get}
\alias{omeka_get}
\title{GET request to an Omeka API}
\usage{
omeka_get(resource, id = NULL, ...)
}
\arguments{
\item{resource}{The resource, such as \code{items} or \code{site}, to
request.}

\item{id}{The (optional) id of a single item to request.}

\item{...}{Additional arguments to be passed to \code{httr::GET()}.}
}
\description{
Use this function to send a GET request to an Omeka site, e.g., to get site
details or items. See the details section or the
\href{http://omeka.readthedocs.org/en/latest/Reference/api/}{Omeka REST API
documentation}.
}
\details{
The resources currently available in Omeka include the following:

  \itemize{ \item \code{collections} \item \code{items} \item \code{files}
  \item \code{item_types} \item \code{elements} \item \code{element_sets}
  \item \code{items} }
}
\references{
See the
  \href{http://omeka.readthedocs.org/en/latest/Reference/api/}{Omeka REST API
  documentation} for more information about the API request you might pass to
  an Omeka site.
}

