% Generated by roxygen2 (4.0.2): do not edit by hand
\name{omeka_key}
\alias{omeka_key}
\title{Get or set the Omeka API key}
\usage{
omeka_key(key = NULL)
}
\arguments{
\item{key}{The Omeka API key to the site that you are using.}
}
\value{
The current Omeka API key, or NULL if none is set.
}
\description{
Pass an Omeka API key to this function to set the Omeka API key for the rest
of your script. Call this function without an argument to get the currently
set endpoint. If you do not set an Omeka API key, then this function returns
NULL and the API will be accessed without passing along a key. You can set
your API key as the \code{OMEKA_KEY} system environment variable.
}
\examples{
omeka_key()
}

