% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{omekaR}
\alias{omekaR}
\alias{omekaR-package}
\title{omekaR: A client for the Omeka API}
\description{
omekaR: A client for the Omeka API
}
\details{
This package provides an API client for the Omeka publishing platform for
scholars, libraries, museums, and archives.
}
\author{
Lincoln Mullen, \email{lincoln@lincolnmullen.com}
}
\keyword{api}
\keyword{omeka}

